chore: separate jumbotron to module
